llm-test-task/requirements.txt
requests~=2.32.3
fastapi~=0.115.11
httpx~=0.28.1
typing~=3.7.4.3
pydantic~=2.10.6
starlette~=0.46.1
pydantic-settings~=2.8.1
python-dotenv~=1.0.1
uvicorn~=0.34.2
flake8
black
isort